pyproject.toml
[tool.poetry]
name = "thumbnail-collector"
version = "0.1.0"
description = ""
authors = ["Takeo Imai <takeo.bono@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
types-requests = "^2.25.6"
google-api-python-client = "^2.21.0"
google-auth = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pycodestyle = "^2.7.0"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
